+++ CHANGELOG.md
@@ -27,6 +27,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - REST Services: Design and Create project
 - REST Service: Accept, reject, archive, start develop, start pilot, start final, finish
 - Project Query and Query Format
-
-### Changed
-- Move project creator email to project
+- REST Services: Create Query and Create CQL DATA Query